# Conflicts:
#	pom.xml

+++ src/main/java/com/binhbkfx02295/cshelpdesk/facebookuser/service/FacebookUserServiceImpl.java
@@ -1,32 +1,46 @@
 package com.binhbkfx02295.cshelpdesk.facebookuser.service;
 
-import com.binhbkfx02295.cshelpdesk.facebookuser.dto.FacebookUserDTO;
-import com.binhbkfx02295.cshelpdesk.facebookuser.dto.FacebookUserDetailDTO;
+import com.binhbkfx02295.cshelpdesk.facebookuser.dto.*;
 import com.binhbkfx02295.cshelpdesk.facebookuser.entity.FacebookUser;
 import com.binhbkfx02295.cshelpdesk.facebookuser.mapper.FacebookUserMapper;
 import com.binhbkfx02295.cshelpdesk.facebookuser.repository.FacebookUserRepository;
 import com.binhbkfx02295.cshelpdesk.util.APIResultSet;
+import com.binhbkfx02295.cshelpdesk.util.PaginationResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class FacebookUserServiceImpl implements FacebookUserService {
-
     private final FacebookUserRepository facebookUserRepository;
     private final FacebookUserMapper mapper;
 
     @Override
-    public APIResultSet<FacebookUserDTO> save(FacebookUserDTO userDTO) {
+    public APIResultSet<FacebookUserDetailDTO> save(FacebookUserDetailDTO userDTO) {
         try {
             FacebookUser user = facebookUserRepository.save(mapper.toEntity(userDTO));
-            APIResultSet<FacebookUserDTO> result = APIResultSet.ok(MSG_SUCCESS, mapper.toDTO(user));
+            APIResultSet<FacebookUserDetailDTO> result = APIResultSet.ok(MSG_SUCCESS, mapper.toDetailDTO(user));
+            log.info(result.getMessage());
+            return result;
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            return APIResultSet.internalError(e.getMessage());
+        }
+    }
+
+    @Override
+    public APIResultSet<FacebookUserDetailDTO> save(FacebookUserFetchDTO userDTO) {
+        try {
+            FacebookUser user = facebookUserRepository.save(mapper.toEntity(userDTO));
+            APIResultSet<FacebookUserDetailDTO> result = APIResultSet.ok(MSG_SUCCESS, mapper.toDetailDTO(user));
             log.info(result.getMessage());
             return result;
         } catch (Exception e) {
@@ -38,7 +52,25 @@ public class FacebookUserServiceImpl implements FacebookUserService {
     @Override
     public APIResultSet<FacebookUserDetailDTO> update(FacebookUserDetailDTO updatedUserDTO) {
         try {
+            FacebookUser entity = facebookUserRepository.get(updatedUserDTO.getFacebookId());
             return Optional.ofNullable(facebookUserRepository.update(mapper.toEntity(updatedUserDTO)))
+                    .map(user -> {
+                        APIResultSet<FacebookUserDetailDTO> result = APIResultSet.ok(String.format("Cập nhật khách hàng ID: %s thành công.", updatedUserDTO.getFacebookId()), mapper.toDetailDTO(user));
+                        log.info(result.getMessage());
+                        return result;
+                    })
+                    .orElseGet(() -> APIResultSet.notFound(MSG_ERROR_USER_NOT_FOUND));
+        } catch (Exception e) {
+            e.printStackTrace();
+            log.info(String.valueOf(e));
+            return APIResultSet.internalError(MSG_ERROR_INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @Override
+    public APIResultSet<FacebookUserDetailDTO> get(String id) {
+        try {
+            return Optional.ofNullable(facebookUserRepository.get(id))
                     .map(user -> {
                         APIResultSet<FacebookUserDetailDTO> result = APIResultSet.ok(MSG_SUCCESS, mapper.toDetailDTO(user));
                         log.info(result.getMessage());
@@ -46,38 +78,13 @@ public class FacebookUserServiceImpl implements FacebookUserService {
                     })
                     .orElseGet(() -> APIResultSet.notFound(MSG_ERROR_USER_NOT_FOUND));
         } catch (Exception e) {
-            return APIResultSet.internalError(MSG_ERROR_INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    @Override
-    public APIResultSet<List<FacebookUserDTO>> search(String name) {
-        try {
-            return APIResultSet.ok(MSG_SUCCESS, facebookUserRepository.search(name).stream().map(mapper::toListDTO).toList());
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            return APIResultSet.internalError(e.getMessage());
-        }
-    }
-
-    @Override
-    public APIResultSet<FacebookUserDTO> get(String id) {
-        try {
-            return Optional.ofNullable(facebookUserRepository.get(id))
-                    .map(user -> {
-                        APIResultSet<FacebookUserDTO> result = APIResultSet.ok(MSG_SUCCESS, mapper.toDTO(user));
-                        log.info(result.getMessage());
-                        return result;
-                    })
-                    .orElseGet(() -> APIResultSet.notFound(MSG_ERROR_USER_NOT_FOUND));
-        } catch (Exception e) {
             log.error(e.getMessage());
             return APIResultSet.internalError(MSG_ERROR_INTERNAL_SERVER_ERROR);
         }
     }
 
     @Override
-    public APIResultSet<List<FacebookUserDTO>> getAll() {
+    public APIResultSet<List<FacebookUserListDTO>> getAll() {
         try {
             return Optional.ofNullable(facebookUserRepository.getAll())
                     .map(users -> APIResultSet.ok(MSG_SUCCESS, users.stream().map(mapper::toListDTO).toList()))
@@ -105,12 +112,49 @@ public class FacebookUserServiceImpl implements FacebookUserService {
     @Override
     public APIResultSet<Void> deleteById(String s) {
         try {
+            if (s == null || s.isEmpty() || s.isBlank()) {
+                return APIResultSet.badRequest("Facebook ID bị thiếu");
+            }
+
+            if (!facebookUserRepository.existsById(s)) {
+                return APIResultSet.badRequest(("Facebook ID không tồn tại"));
+            }
             facebookUserRepository.deleteById(s);
-            log.info("Delete facebookuser {} OK", s);
-            return APIResultSet.ok();
+            log.info("Xóa Khách hàng ID: {} thành công.", s);
+            return APIResultSet.ok(String.format("Xóa Khách hàng ID: %s thành công.", s), null);
         } catch (Exception e) {
             log.error(e.getMessage());
             return APIResultSet.internalError();
         }
     }
+
+    @Override
+    public APIResultSet<PaginationResponse<FacebookUserDetailDTO>> searchUsers(FacebookUserSearchCriteria criteria, Pageable pageable) {
+        try {
+            Map<String, Object> queryResult = facebookUserRepository.search(criteria, pageable);
+            List<FacebookUserDetailDTO> content = ((List<FacebookUser>)queryResult.get("content")).stream().map(mapper::toDetailDTO).toList();
+            PaginationResponse<FacebookUserDetailDTO> response = new PaginationResponse<>();
+            response.setContent(content);
+            response.setTotalElements((long)queryResult.get("totalElements"));
+            response.setPage(pageable.getPageNumber());
+            response.setSize(pageable.getPageSize());
+            response.setTotalPages((int) Math.ceil((double) response.getTotalElements() / response.getSize()));
+            return APIResultSet.ok("search facebook user thanh cong", response);
+        } catch (Exception e) {
+            log.info(Arrays.toString(e.getStackTrace()));
+            return APIResultSet.internalError();
+        }
+    }
+
+    public List<FacebookUserExportDTO> exportSearchUsers(FacebookUserSearchCriteria criteria, Pageable pageable) {
+        try {
+            Map<String, Object> queryResult = facebookUserRepository.search(criteria, pageable);
+            List<FacebookUserExportDTO> content = ((List<FacebookUser>)queryResult.get("content")).stream().map(mapper::toExportDTO).toList();
+            return content;
+        } catch (Exception e) {
+            log.info(Arrays.toString(e.getStackTrace()));
+            return null;
+        }
+    }
+
 }
